backend/requirements.txt
fastapi
uvicorn
sqlalchemy
httpx
python-jose
passlib[bcrypt]
